telegram_service/requirements.txt
python-telegram-bot>=20.0
pandas>=1.5.0
